move toolbar component, add functionality of toolbar component, add drawer for filters, separate filterSearchDropdown component
